SumsofDigits
When N natural numbers are entered, find the sum of the digits of each natural number, and write a program that prints out the natural number with the maximum sum of the digits. Be sure to write a program has  function def digit_sum(x) that takes the sum of the digits of each natural number.