Reversed Prime Number
When N natural numbers are entered, flip over each natural number and write a program that outputs them if the number is a prime number. For example, if 32 is reversed, 23 is a prime number. Then it prints 23. However, if 910 is reversed, it should be digitized as 19. Ignore the zeroes in a row from the first digit. It must write and program the flip function def reverse(x) and the function def isPrime(x) to determine whether it is a prime number.