Dice Game
Imagine a game in which three dice with eyes from 1 to 6 are awarded according to the following rules.
Rule (1): If you have three identical eyes, you will receive a prize of 10,000 + (same eyes) * 1,000.
Rule (2): If there are only two identical eyes, you will receive a prize of 1,000 + (same eyes) * 100.
Rule (3): If all different eyes come out, you will receive a prize of 100  * (the largest of them).
For example, given three eyes 3, 3 and 6, the prize money will be calculated as 1,000+3*100 and will be awarded 1,300.
If the three eyes are given 2, 2 and 2, they will be calculated as 10,000+2*1,000 and will be paid 12,000.
If three eyes are given 6, 2 and 5, the largest value is 6, so it is calculated as 6*100, and the prize money is 600.
When N people participate in the dice game, write a program that prints out the prize money of the person who receives the most prize money.