Score Calculation
An OX problem is a problem that has two correct or wrong answers. In tests made up of multiple OX problems, the following points were calculated to give additional points for consecutive correct answers. If question 1 is correct, calculate it as 1 point. For the previous question, the first question with the correct answer is calculated as one point. Furthermore, if the answer to the question is correct in succession, the second problem is calculated as 2 points, the third problem as 3 points, ..., and the K-th problem as K points. The wrong problem is calculated with zero points.