Substract only numbers
Given a string of letters and numbers, only numbers are extracted and natural numbers are made in that order. It prints out the number of natural number and the divisor of natural number. If you extract only numbers from "t0e0a1c2h0er", it's 0, 0, 1, 2, and if you make a natural number, it's 120. That is, the first digit 0 is ignored when naturalizing. The natural number extracted does not exceed 100,000,000.

input
The first line is given a string mixed with numbers. String length is not more than 50.

output
Print natural numbers on the first line, and the number of divisors on the second line.