CardReversePlacement
Twenty cards with numbers written one by one from 1 to 20 are placed in ascending order.The location of each card is indicated by numbers from 1 to 20, as shown on the card. You now reposition the card with the following rules: given the interval [a, b] (but 1 <= a <= b <= 20), the cards from positions a to b are placed in the current reverse order. For example, if the order in which the cards are placed is the same as in the picture above, and given the interval [5, 10], cards 5, 6, 7, 8, 9, 10 are reversed from position 5 to position 10. If sections [9, 13] are given again, cards 6, 5, 11, 12, 13 from positions 9 to 13 are reversed to 13, 12, 11, 5, 6. Given 10 intervals for 20 cards in ascending order, write a program that successively processes the flip-flops according to the rules in the order of the given interval and then obtains the placement of the last cards.

input
A total of 10 lines are given 10 intervals per line. The i-th row is given the starting position ai of the i-th section and the ending position bi in turn. The two values range from 1 <= ai <= bi <= 20.

Output
For cards placed in ascending order from 1 to 20, the batch of the last cards is printed out in a single line when flipped in the order of the ten sections given as input.