CombineList
If you are given two lists sorted in ascending order, create a program that combines the two lists in ascending order and outputs them.

input
The first line is given the size N (1<=N<=100) of the first list.
In the second line, N list elements are given in ascending order.
The third line is given the size M (1<=M<=100) of the second list.
In the fourth line, M list elements are given in ascending order.
The elements in each list do not exceed the size of the int variable.

Output
Outputs a list sorted in ascending order.